Fixed annihilation field/singularity affecting wrong maps
Right now, those 2 abilities use Find.CurrentMap, which is the map the player is on - not the abilities. This allowed those abilities to affect different maps instead of the one they were cast on by just switching the map the player is on.

This fixes this issue by using the actual map they're located on.